+++ R/data.R
@@ -65,3 +65,12 @@
 #' Derived from the `adsl` and `vs` datasets using `{admiral}` (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_advs.R})
 #'
 "admiral_advs"
+
+#' Pharmacokinetics Parameters Analysis Dataset
+#'
+#' An example vital signs analysis dataset
+#'
+#' @source
+#' Derived from the `adsl` and `pp` datasets using `{admiral}` (\url{https://github.com/Roche-GSK/admiral/blob/main/inst/templates/ad_adpp.R})
+#'
+"admiral_adpp"