+++ inst/templates/ad_adae.R
@@ -2,7 +2,7 @@
 #
 # Label: Adverse Event Analysis Dataset
 #
-# Input: ae, adsl, suppae, ex_single
+# Input: ae, adsl, ex_single
 library(admiral)
 library(admiraltest) # Contains example datasets from the CDISC pilot project
 library(dplyr)
@@ -15,7 +15,6 @@ library(lubridate)
 # For illustration purposes read in admiral test data
 
 data("admiral_ae")
-data("admiral_suppae")
 data("admiral_adsl")
 data("ex_single")
 
@@ -24,7 +23,6 @@ ae <- admiral_ae
 suppae <- admiral_suppae
 
 ae <- convert_blanks_to_na(ae)
-suppae <- convert_blanks_to_na(suppae)
 ex <- convert_blanks_to_na(ex_single)
 
 
@@ -34,8 +32,6 @@ ex <- convert_blanks_to_na(ex_single)
 adsl_vars <- vars(TRTSDT, TRTEDT, DTHDT, EOSDT)
 
 adae <- ae %>%
-  # join supplementary qualifier variables
-  derive_vars_suppqual(suppae) %>%
   # join adsl to ae
   derive_vars_merged(
     dataset_add = adsl,